# docs/swagger.yaml
basePath: /
definitions:
...
  model.Direction:
    enum:
    - asc
    - desc
    type: string
    x-enum-varnames:
    - Asc
    - Desc
...
  model.QueriesRequestElement:
    properties:
      columns:
        items:
          $ref: '#/definitions/model.QueriesRequestElementColumn'
        type: array
      deviceGroupId:
        type: string
      deviceId:
        type: string
      exportId:
        type: string
      filters:
        items:
          $ref: '#/definitions/model.QueriesRequestElementFilter'
        type: array
      groupTime:
        type: string
      limit:
        type: integer
      locationId:
        type: string
      orderColumnIndex:
        type: integer
      orderDirection:
        $ref: '#/definitions/model.Direction'
      serviceId:
        type: string
      time:
        $ref: '#/definitions/model.QueriesRequestElementTime'
    type: object
  model.QueriesRequestElementColumn:
    properties:
      conceptId:
        type: string
      criteria:
        $ref: '#/definitions/models.DeviceGroupFilterCriteria'
      groupType:
        type: string
      math:
        type: string
      name:
        type: string
      sourceCharacteristicId:
        type: string
      targetCharacteristicId:
        type: string
    type: object
  model.QueriesRequestElementFilter:
    properties:
      column:
        type: string
      math:
        type: string
      type:
        type: string
      value: {}
    type: object
  model.QueriesRequestElementTime:
    properties:
      ahead:
        type: string
      end:
        type: string
      last:
        type: string
      start:
        type: string
    type: object
  model.QueriesV2ResponseElement:
    properties:
      columnNames:
        items:
          type: string
        type: array
      data:
        items:
          items:
            items: {}
            type: array
          type: array
        type: array
      deviceId:
        type: string
      exportId:
        type: string
      requestIndex:
        type: integer
      serviceId:
        type: string
    type: object
...
  models.DeviceGroupFilterCriteria:
    properties:
      aspect_id:
        type: string
      device_class_id:
        type: string
      function_id:
        type: string
      interaction:
        $ref: '#/definitions/models.Interaction'
    type: object
  models.Interaction:
    enum:
    - event
    - request
    - event+request
    type: string
    x-enum-varnames:
    - EVENT
    - REQUEST
    - EVENT_AND_REQUEST
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Timescale Wrapper API
  version: "0.1"
paths:
...
  /queries/v2:
    post:
      consumes:
      - application/json
      parameters:
      - description: requested values
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/model.QueriesRequestElement'
          type: array
      - description: specifies output format. Use per_query (default) for a 3D array
          or table for a 2D array with merged timestamps
        in: query
        name: format
        type: string
      - description: Column to order values by (includes time column). Only works
          in format table.
        in: query
        name: order_column_index
        type: string
      - description: Direction to order values by. Allowed are 'asc' and 'desc'. Only
          works in format table.
        in: query
        name: order_direction
        type: string
      - description: 'Textual representation of the date ''Mon Jan 2 15:04:05 -0700
          MST 2006''. Example: 2006-01-02T15:04:05.000Z07:00 would format timestamps
          as rfc3339 with ms precision. Find details here: https://golang.org/pkg/time/#Time.Format'
        in: query
        name: time_format
        type: string
      - description: Used to automatically select the clostest location on a multivalued
          import export. Only works with exportId set to an export of an import. User
          needs read access to the import type.
        in: query
        name: locate_lat
        type: string
      - description: Used to automatically select the clostest location on a multivalued
          import export. Only works with exportId set to an export of an import. User
          needs read access to the import type.
        in: query
        name: locate_lon
        type: string
      - description: Calculate aggregations with the specified timezone instead of
          the default device timezone. Might increase calculation complexity and response
          time.
        in: query
        name: force_tz
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: requestIndex allows to match response and request elements
            (topmost array in request). If a device group is requested, each device
            will return its own time series. If multiple columns are requested, each
            will be return as a time series within the data field. If a criteria is
            selected and multiple paths match the criteria, all matching values will
            be part of the time series.
          schema:
            items:
              $ref: '#/definitions/model.QueriesV2ResponseElement'
            type: array
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: last-values
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"